requirements.txt
anthropic
hypothesis
pytest
python-dotenv
